Rename variations declaration form provider to package playback.DeclarationFormProvider
