+++ app/forms/playback/DeclarationFormProvider.scala
@@ -14,15 +14,16 @@
  * limitations under the License.
  */
 
-package forms
+package forms.playback
 
+import forms.Validation
 import forms.mappings.Mappings
 import javax.inject.Inject
 import models.{DeclarationChangesNoChanges, FullName}
 import play.api.data.Form
 import play.api.data.Forms.{mapping, optional}
 
-class DeclarationChangesNoChangesFormProvider @Inject() extends Mappings {
+class DeclarationFormProvider @Inject() extends Mappings {
 
   def apply(): Form[DeclarationChangesNoChanges] =
   Form(
